# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для расчёта нагрева компонентов",
        "title": "Heating API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/components": {
            "get": {
                "description": "Возвращает все компоненты или фильтрует по названию",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "components"
                ],
                "summary": "Получить список компонентов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название компонента для поиска",
                        "name": "Title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список компонентов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/serializer.ComponentJSON"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает новый компонент (с возможной загрузкой изображения/видео)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "components"
                ],
                "summary": "Создать компонент",
                "parameters": [
                    {
                        "description": "Данные нового компонента",
                        "name": "component",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/serializer.ComponentJSON"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Созданный компонент",
                        "schema": {
                            "$ref": "#/definitions/serializer.ComponentJSON"
                        }
                    },
                    "400": {
                        "description": "Неверные данные",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/components/{id}": {
            "get": {
                "description": "Возвращает информацию о компоненте по идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "components"
                ],
                "summary": "Получить компонент по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID компонента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные компонента",
                        "schema": {
                            "$ref": "#/definitions/serializer.ComponentJSON"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Компонент не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "serializer.ComponentJSON": {
            "type": "object",
            "properties": {
                "component_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "is_deleted": {
                    "type": "boolean"
                },
                "photo_url": {
                    "type": "string"
                },
                "thermal_resistance": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                },
                "video": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}